requirements.txt
streamlit
openai
openpyxl
httpx
google-generativeai>=0.8.3
